+++ examples/devnet-deploy-paymaster/src/reaper/types.rs
@@ -1,0 +1,49 @@
+use std::time::Duration;
+
+use solana_sdk::pubkey::Pubkey;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Loader {
+    V3,
+    V4,
+}
+
+#[derive(Debug, Clone)]
+pub struct OwnedProgram {
+    pub loader: Loader,
+    pub program: Pubkey,
+    /// `None` for v4 — state and code share one account.
+    pub program_data: Option<Pubkey>,
+    pub last_state_slot: u64,
+}
+
+#[derive(Debug, Clone)]
+pub struct ReaperConfig {
+    pub fee_payer: Pubkey,
+    pub threshold: Duration,
+    pub dry_run: bool,
+    pub max_closes: Option<usize>,
+}
+
+#[derive(Debug, Default)]
+pub struct ReaperReport {
+    pub discovered: usize,
+    pub skipped_recent: usize,
+    pub closed: Vec<ClosedProgram>,
+    pub failed: Vec<FailedClose>,
+}
+
+#[derive(Debug)]
+pub struct ClosedProgram {
+    pub program: Pubkey,
+    pub loader: Loader,
+    pub signature: String,
+    pub reclaimed_lamports: u64,
+}
+
+#[derive(Debug)]
+pub struct FailedClose {
+    pub program: Pubkey,
+    pub loader: Loader,
+    pub error: String,
+}
